add generated help to all projects (#253)
* add generated help to all projects

* fix

* Add remaining projects to list

* Add troubleshoot project to list

+++ build/update-attribution-files/make_attribution.sh
@@ -32,11 +32,13 @@ function build::attribution::generate(){
         export RELEASE_BRANCH="$1"
     fi
     make -C $PROJECT_ROOT binaries attribution checksums
-    for summary in $PROJECT_ROOT/_output/**/summary.txt; do
-        sed -i "s/+.*=/ =/g" $summary
-        awk -F" =\> " '{ count[$1]+=$2} END { for (item in count) printf("%s => %d\n", item, count[item]) }' \
-            $summary _output/total_summary.txt | sort > _output/total_summary.tmp && mv _output/total_summary.tmp _output/total_summary.txt
-    done
+    if [ -f $PROJECT_ROOT/_output/**/summary.txt ]; then
+        for summary in $PROJECT_ROOT/_output/**/summary.txt; do
+            sed -i "s/+.*=/ =/g" $summary
+            awk -F" =\> " '{ count[$1]+=$2} END { for (item in count) printf("%s => %d\n", item, count[item]) }' \
+                $summary _output/total_summary.txt | sort > _output/total_summary.tmp && mv _output/total_summary.tmp _output/total_summary.txt
+        done
+    fi
     make -C $PROJECT_ROOT clean 
 }
